requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlmodel==0.0.22
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
redis==5.2.1
alembic==1.14.0
PyJWT==2.10.1
httpx==0.28.1

